# v7: evaluate any rule at most once