This optimization means that complex rules will only be evaluated once
for any given config/user combination - the previous result is used when
possible.

We also make it clear that a User value is intended to be unchanging,
and fix an issue with int-values in percentage rules where they
could return a float64 instead of an int.

Any values that can be computed when the configuration is parsed
are stored then; any that cannot are stored in a slice, created on demand,
that's only as large as it needs to be, and only on demand, which means
that in the common case where the only flags inspected have no rollout
rules, this will be quite efficient.

As a special case when the configuration is parsed, we also compute
the snapshot for the no-user case. This is particularly helpful for long-running
goroutines that aren't servicing a user request because in that case
there's generally no good way to keep a snapshot around for the duration
of a request, and hence the `Client.Snapshot` cost is dominant.
Also, having a single snapshot instance for this case means the cached
calculated values can be shared between all goroutines that use `Client.Snapshot(nil)`.
This improvement is measured by the `Get/no-user/get-and-make` benchmark.
I've also included a benchmark for the likely-common case of getting a value
from a snapshot for a flag with no rollout rules.

The benchmarks show modest cost in some places, but very significant improvement
in others.

```
name                                   old time/op    new time/op    delta
Get/one-of/get-and-make-8                 166ns ± 1%     245ns ± 1%   +47.59%  (p=0.008 n=5+5)
Get/one-of/get-only-8                    64.4ns ± 1%     7.5ns ± 3%   -88.43%  (p=0.008 n=5+5)
Get/less-than-with-int/get-and-make-8     122ns ± 1%     201ns ± 1%   +65.04%  (p=0.008 n=5+5)
Get/less-than-with-int/get-only-8        22.8ns ± 1%     7.4ns ± 0%   -67.53%  (p=0.008 n=5+5)
Get/with-percentage/get-and-make-8        417ns ± 0%     492ns ± 0%   +17.97%  (p=0.008 n=5+5)
Get/with-percentage/get-only-8            312ns ± 1%       7ns ± 2%   -97.61%  (p=0.008 n=5+5)
Get/no-rules/get-and-make-8               106ns ± 0%     118ns ± 0%   +11.91%  (p=0.008 n=5+5)
Get/no-rules/get-only-8                  14.7ns ± 5%     4.4ns ± 3%   -70.15%  (p=0.008 n=5+5)
Get/no-user/get-and-make-8               91.7ns ± 0%    10.6ns ± 0%   -88.41%  (p=0.008 n=5+5)
Get/no-user/get-only-8                   8.03ns ± 3%    7.38ns ± 1%    -8.11%  (p=0.008 n=5+5)
NewSnapshot-8                            8.06µs ± 1%    8.01µs ± 1%      ~     (p=0.151 n=5+5)

name                                   old alloc/op   new alloc/op   delta
Get/one-of/get-and-make-8                 96.0B ± 0%    180.0B ± 0%   +87.50%  (p=0.008 n=5+5)
Get/one-of/get-only-8                     0.00B          0.00B           ~     (all equal)
Get/less-than-with-int/get-and-make-8     96.0B ± 0%    180.0B ± 0%   +87.50%  (p=0.008 n=5+5)
Get/less-than-with-int/get-only-8         0.00B          0.00B           ~     (all equal)
Get/with-percentage/get-and-make-8         168B ± 0%      252B ± 0%   +50.00%  (p=0.008 n=5+5)
Get/with-percentage/get-only-8            72.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)
Get/no-rules/get-and-make-8               96.0B ± 0%    176.0B ± 0%   +83.33%  (p=0.008 n=5+5)
Get/no-rules/get-only-8                   0.00B          0.00B           ~     (all equal)
Get/no-user/get-and-make-8                96.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)
Get/no-user/get-only-8                    0.00B          0.00B           ~     (all equal)
NewSnapshot-8                            4.63kB ± 0%    5.10kB ± 0%   +10.02%  (p=0.008 n=5+5)

name                                   old allocs/op  new allocs/op  delta
Get/one-of/get-and-make-8                  1.00 ± 0%      2.00 ± 0%  +100.00%  (p=0.008 n=5+5)
Get/one-of/get-only-8                      0.00           0.00           ~     (all equal)
Get/less-than-with-int/get-and-make-8      1.00 ± 0%      2.00 ± 0%  +100.00%  (p=0.008 n=5+5)
Get/less-than-with-int/get-only-8          0.00           0.00           ~     (all equal)
Get/with-percentage/get-and-make-8         3.00 ± 0%      4.00 ± 0%   +33.33%  (p=0.008 n=5+5)
Get/with-percentage/get-only-8             2.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
Get/no-rules/get-and-make-8                1.00 ± 0%      1.00 ± 0%      ~     (all equal)
Get/no-rules/get-only-8                    0.00           0.00           ~     (all equal)
Get/no-user/get-and-make-8                 1.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
Get/no-user/get-only-8                     0.00           0.00           ~     (all equal)
NewSnapshot-8                              6.00 ± 0%      6.00 ± 0%      ~     (all equal)
```

### Requirement checklist (only if applicable)

- [x] I have covered the applied changes with automated tests.
- [x] I have executed the full automated test set against my changes.